lab4/api/requirements.txt
fastapi
uvicorn
kafka-python
psycopg2-binary
redis
pydantic
